refactoring and removing mathamtical details from documentation
