--- src/adaptive_quadrature/mod.rs
@@ -1,12 +1,0 @@
-//! Adaptive Methods
-//!
-//! Adaptive methods are used to estimate the integral of a function on a closed and
-//! bounded interval. The basic idea is to take smaller subintervals in regions where the
-//! function changes significantly and larger subintervals in regions where the function
-//! doesn't change significantly. What is significant depends on the quadrature technique
-//! involved, for Simpson's rule, if the function is nearly cubic (or quadratic, or linear, or
-//! constant) then the function doesn't change significantly from being cubic and a large
-//! interval can be used.
-//!
-
-pub mod simpson;
