+++ src/utils/checkers.rs
@@ -1,4 +1,4 @@
-use num::{Float, Unsigned};
+use num::{Float, Unsigned, Zero};
 
 /// Checks integral arguments for Newton-Codes methods
 ///
@@ -16,5 +16,14 @@ pub fn check_newton_method_args<F: Float, U: Unsigned>(a: F, b: F, n: U) {
 
     if a > b {
         panic!("a must be strictly less than b");
+    }
+}
+
+/// Checks integral arguments for Gauss-Laguerre rule
+///
+/// * `n` - number of steps.
+pub fn check_gauss_rule_args(n: usize) {
+    if n.is_zero() {
+        panic!("number of steps can't be zero");
     }
 }
